test/regression/reftest/vendor/blink/fast/dom/Text/previous-element-sibling.html

<!DOCTYPE html>
<script src="../../../resources/wpt-test-pre.js"></script>
<script src="../../../resources/js-test-pre.js"></script>
<script>

description('This tests the NonDocumentTypeChildNode interface previousElementSibling property on a Text node.');

var element = document.createElement('div');
var a = element.appendChild(new Text('a'));
shouldBeNull('a.previousElementSibling');
var b = element.insertBefore(document.createElement('b'), a);
shouldBe('a.previousElementSibling', 'b');
var c = element.insertBefore(new Comment('c'), a);
shouldBe('a.previousElementSibling', 'b');

</script>
<script src="../../../resources/js-test-post.js"></script>
<script src="../../../resources/wpt-test-post.js"></script>
